feat: add authentication

+++ site/pages/api/login.api.ts
@@ -1,5 +1,13 @@
+import { NotAuthorizedException } from "@aws-sdk/client-cognito-identity-provider";
 import { NextApiRequest, NextApiResponse } from "next";
-import { COOKIES_LOGIN_ERRORS, DASHBOARD_PAGE_PATH, LOGIN_PAGE_PATH } from "../../constants";
+import {
+    COOKIES_ID_TOKEN,
+    COOKIES_LOGIN_ERRORS,
+    COOKIES_REFRESH_TOKEN,
+    DASHBOARD_PAGE_PATH,
+    LOGIN_PAGE_PATH,
+} from "../../constants";
+import { initiateAuth } from "../../data/cognito";
 import { loginSchema } from "../../schemas/login.schema";
 import { flattenZodErrors } from "../../utils";
 import {
@@ -8,8 +16,9 @@ import {
     redirectTo,
     destroyCookieOnResponseObject,
 } from "../../utils/apiUtils";
+import logger from "../../utils/logger";
 
-const login = (req: NextApiRequest, res: NextApiResponse) => {
+const login = async (req: NextApiRequest, res: NextApiResponse) => {
     try {
         const validatedBody = loginSchema.safeParse(req.body);
         if (!validatedBody.success) {
@@ -26,10 +35,47 @@ const login = (req: NextApiRequest, res: NextApiResponse) => {
             return;
         }
 
+        const { email, password } = validatedBody.data;
+
+        const authResponse = await initiateAuth(email, password);
+
+        if (!authResponse.AuthenticationResult?.IdToken || !authResponse.AuthenticationResult?.RefreshToken) {
+            throw new Error("Invalid auth response");
+        }
+
+        const { IdToken: idToken, RefreshToken: refreshToken } = authResponse.AuthenticationResult;
+
+        setCookieOnResponseObject(COOKIES_ID_TOKEN, idToken, res);
+        setCookieOnResponseObject(COOKIES_REFRESH_TOKEN, refreshToken, res);
+
+        logger.info("", {
+            context: "api.login",
+            message: "login successful",
+        });
+
         destroyCookieOnResponseObject(COOKIES_LOGIN_ERRORS, res);
         redirectTo(res, DASHBOARD_PAGE_PATH);
         return;
     } catch (e) {
+        if (e instanceof NotAuthorizedException) {
+            setCookieOnResponseObject(
+                COOKIES_LOGIN_ERRORS,
+                JSON.stringify({
+                    inputs: req.body as object,
+                    errors: [
+                        {
+                            errorMessage: "Incorrect username or password",
+                            id: "",
+                        },
+                    ],
+                }),
+                res,
+            );
+
+            redirectTo(res, LOGIN_PAGE_PATH);
+            return;
+        }
+
         if (e instanceof Error) {
             const message = "There was a problem during login.";
             redirectToError(res, message, "api.login", e);
